backend/pyproject.toml
[project]
name = "ama-backend"
version = "0.1.0"
description = "AMA backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "google-genai>=1.66.0",
    "pillow>=12.1.1",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "rembg[cpu]>=2.0.0",
]
